requirements/tests.txt
-r base.txt

flake8==3.7.9
isort==4.3.21
django-debug-toolbar==2.2
nplusone==1.0.0
